requirements.txt
flask==2.2.3
flask-cors==3.0.10
Pillow>=10.0.0
requests==2.27.1
python-dotenv==0.19.2
werkzeug==2.2.3